fix : 좀 더 추상화
